requirements.txt
amqp==1.4.9
anyjson==0.3.3
arrow==0.12.1
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
billiard==3.3.0.23
blessed==1.14.2
cachetools==2.1.0
celery==3.1.24
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
diff-match-patch==20121119
Django==1.8.5
django-admin-rangefilter==0.3.6
django-allauth==0.26.1
django-appconf==1.0.2
django-audit-log==0.7.0
django-celery==3.1.17
django-common-helpers==0.9.1
django-cron==0.4.6
django-datatables-view==1.14.0
django-daterange-filter==1.3.0
django-debug-toolbar==1.8
django-easy-select2==1.3.3
-e git://github.com/mvpdev/django-eav.git@86098dc1cd602be8cf2e7fd8fbbc34a6e6b2dd53#egg=django_eav
django-filter==0.15.3
django-imagekit==3.3
django-impersonate==1.3
django-import-export==0.5.1
django-multiselectfield==0.1.8
django-notifier==0.7
django-picklefield==1.0.0
django-push-notifications==1.4.1
django-q==0.8.1
django-redis==4.6.0
django-redis-cache==1.7.1
django-rest-auth==0.8.1
django-versatileimagefield==1.9
djangorestframework==3.3.2
drf-nested-routers==0.11.1
drf-tracking==1.4.0
drfdocs==0.0.11
elasticsearch==6.1.1
elasticsearch-dsl==6.1.0
enum==0.4.6
enum34==1.1.6
et-xmlfile==1.0.1
future==0.16.0
gevent==1.3.2.post0
google-api-python-client==1.7.3
google-auth==1.5.0
google-auth-httplib2==0.0.3
greenlet==0.4.13
grequests==0.3.0
httplib2==0.11.3
idna==2.6
ipaddress==1.0.22
ipython==5.7.0
ipython-genutils==0.2.0
jdcal==1.4
kombu==3.0.37
MySQL-python==1.2.5
oauth2client==4.1.2
oauthlib==2.1.0
odfpy==1.3.6
openpyxl==2.5.3
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.4
pilkit==2.0
Pillow==5.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycrypto==2.6.1
pycurl==7.43.0
Pygments==2.2.0
python-dateutil==2.7.3
python-openid==2.2.5
pytz==2018.4
PyYAML==3.12
redis==2.10.5
requests==2.13.0
requests-oauthlib==0.8.0
rsa==3.4.2
scandir==1.7
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
tablib==0.12.1
traitlets==4.3.2
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
xlrd==1.1.0
xlwt==1.3.0
